Add rebar support.
Add support for dependent projects and a rebar toolchain.

Dependent projects are specified via the "DEPENDENCIES" argument.
When provided its is expected to be a list of tuples of (name, url)
where url is a valid checkout url for hte project. At the moment git
is the assumed repository type.

Dependencies are cloned if nonexistant during configure, otherwise they
are updated.

They are compiled/instlled/doc generated before regular apps.