New config class
Class takes an input source and parses it into config.

Treats all keys in a case insensitive manner.

Proplists are merged into dictionaries.
When a key is specified more than once all values are stored
at that key as a list. Order is not gauranteed.

If a dicitonary of types is provided, values assigned the keys
contained within will be processed against the function contained
within. Ergo types={a:int) config[a] = "1" config[a] -> 1